kootumb_API_Backend/requirements.txt
# ============================================================
# KOOTUMB BACKEND - COMPLETE REQUIREMENTS
# ============================================================
# All packages needed for full functionality

# Core Django Framework (ESSENTIAL)
Django==2.2.28
djangorestframework==3.12.4
setuptools>=40.0.0  # Required for Python 3.12 compatibility
django-replicated>=2.5.0  # Required by settings.py

# Database & ORM
mysqlclient==2.2.0  # MySQL database adapter for Django

# Authentication & Security (ESSENTIAL)
PyJWT==1.7.1

# File Processing & Media (ESSENTIAL)
Pillow==8.4.0
pilkit==2.0
django-imagekit==4.0.2

# AWS Cloud Storage
boto3==1.20.0
django-storages==1.11.1

# Background Tasks & Queue Management (ESSENTIAL)
celery==5.2.7
django-rq==2.5.1
redis==4.3.4
django-redis==4.12.1  # Redis cache backend (compatible with Django 2.2)

# API Extensions
rest-framework-generic-relations==2.0.0

# Core Utilities (ESSENTIAL)
python-dotenv==0.19.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1

# Development & Debug Tools
django-extensions==3.1.5
django-media-fixtures>=1.0.0  # Media fixtures for development

# Internationalization & Translation
django-modeltranslation==0.17.3

# Media & Content Processing
ffmpy==0.3.0
webpreview==1.6.0

# Model Utilities & Helpers
django-ordered-model==3.6.0
django-model-utils==4.2.0

# API Pagination
django-cursor-pagination==0.1.4

# Testing & Quality Assurance
faker==15.3.4
# django-nose==1.4.7  # Commented - not compatible with Python 3.12

# Content Processing & Analysis
beautifulsoup4==4.11.1
colormath==3.0.0
natsort==8.2.0
langdetect==1.0.9
python-magic==0.4.27
tldextract>=5.0.0  # URL domain extraction

# Configuration & Monitoring
pyyaml==5.4.1
sentry-sdk==1.9.10
django-positions>=0.6.0
PyYAML>=6.0.2
django-proxy>=1.3.0
shutilwhich>=1.1.0
spectra>=0.1.0
url-normalize>=2.2.1
urlextract>=1.9.0
onesignal==0.1.3
onesignal-client==0.0.2>=2.0.0  # Push notifications

# ASGI Server (ESSENTIAL for production)
uvicorn[standard]>=0.35.0  # High-performance ASGI server


#mysqlclient